#67 Fixed overflow problem and use weighted (area) average to calculate surface normals

+++ src/models/ModelData.h
@@ -10,11 +10,11 @@ public:
 	ModelData();
 	~ModelData();
 
-    std::vector<GLushort> elements;
+    std::vector<int> elements;
     std::vector<float> bcCoords;
     std::vector<int> closestParticles;
 
     void clear();
 
-    void addVertices(std::vector<short> &e, std::vector<float[3]> &bc, std::vector<int[3]> &particles, ModelData &data);
+    void addVertices(std::vector<int> &e, std::vector<float[3]> &bc, std::vector<int[3]> &particles, ModelData &data);
 };
